Tests running, Dockerfile adjusted, absolute imports added.
