+++ .env
@@ -1,0 +1,2 @@
+AWS_ACCESS_KEY_ID=AKIAVLPJWCYQOHNFVRXT
+AWS_SECRET_ACCESS_KEY=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
