--- app/Dockerfile
@@ -1,25 +1,0 @@
-# 
-FROM python:3.9
-
-# 
-WORKDIR /app
-
-# 
-COPY ./requirements.txt /app/requirements.txt
-
-# 
-RUN pip install --no-cache-dir --upgrade -r /app/requirements.txt
-# RUN pip install --no-cache-dir -r requirements.txt
-
-# 
-COPY ./ /app
-
-#
-# ENV PYTHONPATH "${PYTHONPATH}:/code/app"
-
-# 
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "80"]
-
-
-
-
