+++ requirements.txt
@@ -3,6 +3,7 @@ pydantic>=1.8.0,<2.0.0
 uvicorn>=0.15.0,<0.16.0
 typing-extensions>=3.7.4.2
 typer>=0.3.2,<0.4.0
+pytest>=6.2.4,<7.0.0
 validators
 pynamodb
 requests
